Rewrite the buffering in MultiThreadedStreamer
Rewrite the buffering in MultiThreadedStreamer to make it more
efficient in case the buffer size is large. Specifically, store the
unused buffer chunks in a stack instead of a queue to improve memory
locality. (The filled buffer chunks still need to be a queue, because
we need to consume them in FIFO order.)

This obsoletes https://github.com/kwiberg/frz/issues/4.