Add support for forked streams
Until now, all streams were "straight", that is, one source streaming
bytes to one sink. This commit adds a second type, "forked" streams,
where one source streams bytes to two sinks. We always stream bytes to
the primary sink as fast as possible, and make a best effort to stream
the same bytes to the secondary sink, but if it can't keep up, we
pause it and stream only to the primary sink. When the source is
exhausted and the primary sink has all the bytes, we let the caller
decide whether to abandon the secondary sink, or rewind the source and
feed the secondary sink all the bytes it didn't receive the first time
around.

The idea is to use this when looking for missing content; given a
candidate file, we stream it to a hasher (primary sink) and a
file (secondary sink); in case the hash value turns out to be what we
hoped, we will already have copied a significant chunk of the file to
where we need it to be (and if it isn't, we can just delete the file).
A future commit will do this.